fixed two problems
PlayArea and DeckBehaviour were starting at different times, causing the
first to not find XML loaded and the second able to. Changed the script
boot order to fix.
There was a second problem, the activeSquares were being instantaited,
but Start() for them was not given enough time  to set the
spriteRenderer to a value before the method was called to activate them.
Changing Start() to Awake() fixed this,