Implimenting body part disablement
started putting in calls to be able to detect health dropping bellow
body part max, then calling a 'disable' of all the squares associated
with it. Only left arm works consistently. Call for checking to disable
body parts is in gamecontroller update(). Might not be the best place.

Need to revise and edit calls and methods, getting too disorganized