implimented a deck and play area for both PC & npc
deck and play area are now prefabs, they can exist without throwing any
errors but the card play still only works for the enemy area. Need to
make the cards work only on the opponent model and for them to work
independtly. This seperation will mostly occur in General and within the
card.